requirements.txt
gradio>=4.44.0
openai>=1.40.0
fal-client>=0.5.0
httpx>=0.27.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0
